feat(idl-gen): Make snapshot tests integrational
