+++ client-gen/src/generator.rs
@@ -718,62 +718,3 @@ impl<'ast> Visitor<'ast> for CallBuilderGenerator {
         self.code.push_str(") -> Vec<u8>");
     }
 }
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_basic_works() {
-        let idl = r"
-            type MyParam = struct {
-                f1: u32,
-                f2: vec str,
-                f3: opt struct { u8, u32 },
-            };
-
-            type MyParam2 = enum {
-                Variant1,
-                Variant2: u32,
-                Variant3: struct { u32 },
-                Variant4: struct { u8, u32 },
-                Variant5: struct { f1: str, f2: vec u8 },
-            };
-
-            service {
-                DoThis: (p1: u32, p2: MyParam) -> u16;
-                DoThat: (p1: struct { u8, u32 }) -> u8;
-            };
-        ";
-
-        let program = sails_idl_parser::ast::parse_idl(idl).expect("parse IDL");
-
-        insta::assert_snapshot!(generate(program, "Basic").unwrap());
-    }
-
-    #[test]
-    fn test_multiple_services() {
-        let idl = r"
-            service {
-                DoThis: (p1: u32, p2: MyParam) -> u16;
-                DoThat: (p1: struct { u8, u32 }) -> u8;
-            };
-
-            service Named {
-                query That: (p1: u32) -> str;
-            };
-        ";
-
-        let program = sails_idl_parser::ast::parse_idl(idl).expect("parse IDL");
-
-        insta::assert_snapshot!(generate(program, "Multiple").unwrap());
-    }
-
-    #[test]
-    fn test_rmrk_works() {
-        let idl = include_str!("../../examples/rmrk/catalog/wasm/rmrk-catalog.idl");
-
-        let program = sails_idl_parser::ast::parse_idl(idl).expect("parse IDL");
-
-        insta::assert_snapshot!(generate(program, "RmrkCatalog").unwrap());
-    }
-}
